actualizacion tipo datos

+++ CodigoFuente/antenas/API/DataSchema/ef_plantillas_evento.cs
@@ -9,7 +9,7 @@ namespace API.DataSchema
         public short id_plantilla { get; set; }
         public string codigo { get; set; }
         public bool activo { get; set; }
-        public short? id_tipo_evento { get; set; }
+        public int? id_tipo_evento { get; set; }
 
         // Nav
         public ICollection<ef_plantilla_tramos>? tramos { get; set; } = new List<ef_plantilla_tramos>();
